salvando foto com anotação mas so a anotação

+++ Quadro/Materia.m
@@ -33,7 +33,16 @@
 
 - (void)deleteMateria {
     if ([[Managerdb sharedManager] opendb]) {
-        [[[Managerdb sharedManager] database] executeUpdate:@"delete from materia where nome=?",self.nome];
+        FMResultSet *rs = [[[Managerdb sharedManager] database] executeQuery:@"select * from materia where nome=?",self.nome];
+        NSString *idMateria = @"0";
+        if ([rs next]) {
+            idMateria =  [NSString stringWithFormat:@"%d", [rs intForColumn:@"idMateria"]];
+            bool deu = [[[Managerdb sharedManager] database] executeUpdate:@"delete from assunto where idMateria=?",idMateria];
+            NSLog(@"%d", deu);
+            deu = [[[Managerdb sharedManager] database] executeUpdate:@"delete from materia where nome=?",self.nome];
+            NSLog(@"%d", deu);
+        }
+        [rs close];
         [[Managerdb sharedManager] closedb];
     }
 }
